+++ release-notes.md
@@ -4,7 +4,11 @@ We'll be updating this document along with each update that is merged into maste
 
 We are now on Testflight :tada:, please forward us an email at hello@dwyl.io to request beta access
 
-## Version 1.1 (WIP Build 13) _Current_
+## Version 1.1 (WIP Build 14) _Current_
+
+- Syncing timer both ways. Users can now see the timer running on the watch if they've started it on the Phone.
+
+## Version 1.1 (WIP Build 13)
 
 - Timer state is reloaded if the user force quits the app and rejoins. :tada: